funziona! non ho cacolato tempo sclo

+++ Lab09_MetroDeParis/src/it/polito/tdp/metrodeparis/model/Linea.java
@@ -28,7 +28,7 @@ public void setNome(String nome) {
 public double getVelocita() {
 	return velocita;
 }
-public void setVelocita(int velocita) {
+public void setVelocita(double velocita) {
 	this.velocita = velocita;
 }
 @Override
